pic2mp4/requirements.txt
ffmpeg-python
pillow
sv_ttk
